Code improvements

+++ app/src/main/java/com/dlminfosoft/sphmobile/model/DataUsageResponse.kt
@@ -9,4 +9,4 @@ data class ResultData(
     val records: ArrayList<RecordsData>, val total: Int, val offset: Int = 0, val limit: Int = 0
 )
 
-data class RecordsData(val _id: Int, val quarter: String, val volume_of_mobile_data: Float)
+data class RecordsData(val _id: Int, val quarter: String, val volume_of_mobile_data: Double)
